# upgrade generate_permute_indices with faster kernel (parallelized)