This PR updates both llama4 and DeepSeekv3 to use the faster parallelized kernel for generate_permute_indices. 
This also adds a unit test for the core kernel in generate_permute_indices.

~~~

Summary for fixed total experts (256), where we mimic various DeepSeek experts across a number of ranks:
Ranks | Experts/Rank | Original (ms) | Optimized (ms) | Block Size | Speedup | Correct
------------------------------------------------------------------------------------------
    1 |          256 |        0.241 |        0.020 |        128 |   11.95x |       1
    4 |           64 |        0.222 |        0.020 |        128 |   10.97x |       1
   16 |           16 |        0.221 |        0.021 |        128 |   10.40x |       1
   32 |            8 |        0.224 |        0.025 |         64 |    8.81x |       1
   64 |            4 |        0.225 |        0.037 |        128 |    6.01x |       1
  128 |            2 |        0.217 |        0.051 |        128 |    4.26x |       1
  ~~~
  
  Testing:
  Verified Llama4 working with new kernel 
  Verified DeepSeekv2 inference working with new kernel
  Verified unit tests passing